% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataM2C.R
\docType{data}
\name{dataM2C}
\alias{dataM2C}
\title{Example network for ncRNA network on miRNA-circRNA}
\usage{
data("dataM2C")
}
\description{
Data example consisting of a matrix of network.
}
\examples{
data("dataM2C")
head(dataM2C[1:100,])
}
