% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataM2L.R
\docType{data}
\name{dataM2L}
\alias{dataM2L}
\title{Example network for ncRNA network on miRNA-lncRNA}
\usage{
data("dataM2L")
}
\description{
Data example consisting of a matrix of network.
}
\examples{
data("dataM2L")
head(dataM2L[1:100,])
}
