% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataM2R.R
\docType{data}
\name{dataM2R}
\alias{dataM2R}
\title{Example network for ncRNA network on miRNA-RPB}
\usage{
data("dataM2R")
}
\description{
Data example consisting of a matrix of network.
}
\examples{
data("dataM2R")
dataM2R[1:100,]
}
