% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataR2C.R
\docType{data}
\name{dataR2C}
\alias{dataR2C}
\title{Example network for ncRNA network on RPB-circRNA}
\usage{
data("dataR2C")
}
\description{
Data example consisting of a matrix of network.
}
\examples{
data("dataR2C")
dataR2C[1:100,]
}
