% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataR2L.R
\docType{data}
\name{dataR2L}
\alias{dataR2L}
\title{Example network for ncRNA network on RBP-lncRNA}
\usage{
data("dataR2L")
}
\description{
Data example consisting of a matrix of network.
}
\examples{
data("dataR2L")
dataR2L[1:100,]
}
