% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IDsymbol.R
\docType{data}
\name{IDsymbol}
\alias{IDsymbol}
\title{Table of the gene identifier and the gene symbol}
\usage{
data("IDsymbol")
}
\description{
Data  consisting of a matrix of the colname of gene_ID and gene_Name
}
\examples{
data("IDsymbol")
IDsymbol[1:100,]
}
